refactor(TemplateLibrary): modularize optional buttons - I62
